Map Preview Component Refactored
