+++ package.json
@@ -75,7 +75,6 @@
     "check-dependency-version-consistency": "^4.1.0",
     "cross-env": "^7.0.3",
     "dotenv-cli": "^7.3.0",
-    "eslint": "^8.47.0",
     "husky": "^8.0.0",
     "lint-staged": "^14.0.1",
     "npm-run-all": "^4.1.5",
@@ -126,6 +125,7 @@
   "devDependencies": {
     "@babel/core": "^7.23.9",
     "@babel/preset-react": "^7.23.3",
-    "babel-loader": "^9.1.3"
+    "babel-loader": "^9.1.3",
+    "eslint": "^8.56.0"
   }
 }
