+++ packages/app/components/map/MapContainer.native.tsx
@@ -4,8 +4,9 @@ import { View, Platform } from 'react-native';
 import { isObjectEmpty } from '../../utils/isObjectEmpty';
 import { defaultShape } from '../../utils/mapFunctions';
 import useTheme from '../../hooks/useTheme';
-import NativeMap from './NativeMap';
+import NativeMap from './NativeMap.native';
 import useCustomStyles from 'app/hooks/useCustomStyles';
+import loadStyles from './MapStyles';
 
 export function MapContainer({ shape }) {
   const { enableDarkMode, enableLightMode, isDark, isLight, currentTheme } =
@@ -30,18 +31,3 @@ export function MapContainer({ shape }) {
 }
 
 export default MapContainer;
-
-const loadStyles = () => ({
-  webContainer: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: 'fit-content',
-  },
-  nativeContainer: {
-    width: '100%',
-    marginBottom: 20,
-    marginBottom: 20,
-    paddingHorizontal: 5,
-  },
-});
