+++ packages/app/components/map/NativeMap.native.tsx
@@ -33,6 +33,7 @@ import { DOMParser } from 'xmldom';
 import { gpx as toGeoJSON } from '@tmcw/togeojson';
 import useCustomStyles from 'app/hooks/useCustomStyles';
 import { useNativeMap } from 'app/hooks/map/useNativeMap';
+import loadStyles from './MapStyles';
 
 Mapbox.setWellKnownTileServer(Platform.OS === 'android' ? 'Mapbox' : 'mapbox');
 Mapbox.setAccessToken(MAPBOX_ACCESS_TOKEN);
@@ -182,7 +183,7 @@ function NativeMap({ shape: shapeProp }) {
               clusterMaxZoomLevel={14}
               style={{ zIndex: 1 }}
             >
-              <Mapbox.LineLayer id="layer1" style={styles.lineLayer} />
+              <Mapbox.LineLayer id="layer1" style={styles.NativeMaplineLayer} />
             </Mapbox.ShapeSource>
             {/* // top location */}
             {shape?.features[0]?.geometry?.coordinates?.length > 0 && (
@@ -342,56 +343,5 @@ function NativeMap({ shape: shapeProp }) {
     </SafeAreaView>
   );
 }
-
-const loadStyles = () => ({
-  page: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  container: {
-    // height: 500,
-    width: '100%',
-    backgroundColor: 'white',
-    marginBottom: 20,
-    marginBottom: 20,
-    paddingHorizontal: 5,
-  },
-  map: {
-    flex: 1,
-  },
-  lineLayer: {
-    lineColor: '#16b22d',
-    lineWidth: 4,
-    lineOpacity: 1,
-  },
-  headerView: {
-    position: 'absolute',
-    marginTop: 20,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-evenly',
-    width: 100,
-  },
-  headerBtnView: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderRadius: 32,
-    backgroundColor: 'white',
-  },
-  button: {
-    position: 'absolute',
-    bottom: 10,
-    right: 10,
-    backgroundColor: theme.colors.primary,
-    borderRadius: 50,
-    width: 45,
-    height: 45,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
 
 export default NativeMap;
